+++ server/database/database-service.js
@@ -3,8 +3,6 @@
  * @module database-service
  */
 
-const mongoose = require('mongoose')
-
 /**
  * Create a query for find function of mongoose.
  * @param {Object} model - mongoose model.
@@ -71,30 +69,9 @@ const findingOne = (model, query, projection = {}, populate = []) => {
 }
 
 /**
- * @param {string} path - name of field that generate the error
- * @param {string} message - custom message
- * @param {string} type - kind of the error (required, duplicated…)
- * @param {string} value - value of path (required, duplicated…)
- * @returns {ValidationError} create a custom error in mongoose style
- */
-const validationError = (path, message, type, value) => {
-  const ValidationError = mongoose.Error.ValidationError
-  const ValidatorError = mongoose.Error.ValidatorError
-  const error = new ValidationError(null)
-  error.errors[path] = new ValidatorError({
-    path,
-    message,
-    type,
-    value
-  })
-  return error
-}
-
-/**
  * Common functions for CRUD operations
  */
 module.exports = Object.freeze({
   finding,
-  findingOne,
-  validationError
+  findingOne
 })