+++ server/error-handler/custom-error.spec.js
@@ -1,25 +1,25 @@
 /* global describe, it */
 /**
- * Test for database module
- * @module database-service.spec
+ * Test for custom-error module
+ * @module custom-error.spec
  */
 
 // Set the env variable to test
 process.env.NODE_ENV = 'test'
 
 const chai = require('chai')
-const dbService = require('./database-service')
+const customError = require('./custom-error')
 
 // Chai styles
 const expect = chai.expect
 
 // Parent block
-describe('database module', () => {
+describe('custom-error module', () => {
   it('it should be create a custom error with mongoose style', () => {
     const path = 'email'
     const message = 'Custom error'
     const kind = 'duplicated'
-    const error = dbService.validationError(path, message, kind, path)
+    const error = customError.validationError(path, message, kind, path)
     expect(error).to.be.a('object')
     expect(error).to.be.have.property('name').eql('ValidationError')
     expect(error).to.be.have.property('errors')